restructed tests
